Add files via upload
The Tourism system works to help people wishing to travel to
a specific country to decide which city should to visit in that
country through the tourist's answers.

Therefor, we will write a set of questions for the tourist about
his interest in this trip and the places wants to visit, that help us to
decide the expected city. The system will be written as an AI code
agent in the prolog program.